Change /etc/ntp.conf to put the driftfile in /var/lib/ntp on Linux.
What I thought was the LCD for all OSes, /etc/ntp.drift, doesn't work
because ntpd does not run as root and it wants to create a tmp file for
the new drift value in the same dir as the driftfile.

So if /var/lib/ntp exists in the image, we modify /etc/ntp.conf to put
the driftfile there. The ntpstart and watchdog scripts should pick this
up just fine as is.